requirements.txt
appdirs==1.4.3
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
cssselect==1.1.0
Django==2.2.6
django-debug-toolbar==2.0
django-summernote==0.8.11.4
fake-useragent==0.1.11
idna==2.8
lxml==4.4.1
parse==1.12.1
Pillow==6.2.0
pyee==6.0.0
pyppeteer==0.0.25
pyquery==1.4.0
python-slugify==3.0.6
pytz==2019.3
requests==2.22.0
requests-html==0.10.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
text-unidecode==1.3
tqdm==4.36.1
urllib3==1.25.6
w3lib==1.21.0
websockets==8.0.2
